lilith/cuda/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cuda"
dynamic = ["version"]
description = "CUDA GPU compute workspace"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "pandas>=2.3.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
notebook = [
    "jupyterlab>=4.0.0",
    "ipykernel>=7.0.0",
    "ipywidgets>=8.0.0",
    "notebook>=7.0.0",
]
science = [
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
]
dl = [
    # torch, torchvision, torchaudio come from the pytorch/pytorch base image
    "transformers>=4.40.0",
    "datasets>=2.20.0",
    "accelerate>=0.30.0",
    "tensorflow>=2.17.0",
    # Frameworks para ativar futuramente:
    # "jax[cuda12]>=0.4.0",   # programação funcional diferenciável (+ flax, optax)
    # "xgboost>=2.0.0",       # boosting para tabular data com GPU
    # "lightgbm>=4.0.0",      # boosting alternativo com GPU
]
dev = [
    "pytest>=9.0.0",
    "ruff>=0.14.0",
    "pip>=25.0",
]

# =============================================================================
# TOOL CONFIGURATION
# =============================================================================

[tool.uv]
system-site-packages = true

[tool.hatch.version]
path = ".env"
pattern = "VERSION=['\"]?(?P<version>[^'\"\\s]+)['\"]?"

[tool.hatch.build]
packages = ["cuda"]

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.ruff]
line-length = 99
src = ["cuda"]
include = ["pyproject.toml", "cuda/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["cuda"]
force-sort-within-sections = true
